Link: converted 'eps' from list to SlotList, introduced foreign channel IDs
In order to prevent race conditions when new channels are created, the IDs used at the one Link's end are not the same as the IDs used at the other end.
From the perspective of a Link, the local channel ID (simply called channelID) is used everywhere except for the headers of the outgoing packets. When sending a packet, the foreign channel ID (called channelIDF), which is the one used by the other end, is used instead. Therefore, the received packets have a local channel ID.
In the Control Channel, there are some special cases, as the IDs appear inside the messages.