Reworked timeouts
SSL sockets need a positive timeout before connecting (blocking). After the handshake, they need a timeout of 0 (non-blocking) when receiving to return quickly if partial SSL records are received. When sending, without addidtional buffering, a positive timeout (blocking) is needed in order to prevent dropping data. The same applies to regular sockets, although it's possible to keep them with timeout even on receive.